Added Fish Movement
Tested fish movement with walls.

Fish are set to kinematic rigidbodies. In order to prevent them from
moving off the boundaries of the playspace, I implemented a function
that checks whether they are about to go off the playspace, and if they
are, then they are to change direction to move away from the edge. At
first, I implemented this function in the OnTriggerEnter2D, however, it
seems as though it is better fit to be implemented in the Update()
function. This is because the trigger function may not occur every
frame, and this can cause the deter function to not be called, and thus
fish can exit the camera frame. This has happened with 25 fish.

Fish movement is determined randomly, and the time spent traveling in a
certain direction is also determined randomly (or psuedorandomly).

Tested on batches of fish, but have yet to test it on > 500 or so fish.