Added Cohesion behavior and steering
Adjust code to allow for easy testing of cohesion behavior.
Thus removed random movement components (with the exception of the
beginning).
Added constants that correspond to Boids algorithm.
Used neighbors within radius as a place holder for the to be
implemented KNN algorithm.
Made sprites rotate in accordance to their direction of travel.
Cleaned up deter and check out of bounds code.
Added shell of flock code (or boid algorithm). The only component
implemented thus far is cohesion behavior.
Added the steer function, which steers a fish towards the position of
interest gradually, in correspondence to how much the STEER_CONST
allows it turn.