Added alignment behavior
The second component of the boids algorithm.
Calculates the average velocity of the neighborhood fish and steers
towards that direction.
Updated the steerTo function to take in the intended velocity of travel
(or the velocity that the function is meant to steer towards), rather
than the target position.
The above change also meant that I needed to change what I passed into
steerTo in the previous cohesion function. Instead of the target
position, I passed in the difference between the target position and
the current position.
Removed distance component of steerTo since it is unnecessary, and an
artifact of a previous misconception of mine of the function.
Updated constants in order to test the alignment function. Seems to be
working A-okay.