Added Separation Behavior
Separation behavior is determined by finding the fish that are within a
certain radius, and traveling in the opposite direction of the fish
that are close enough.
Adjusted Boids weights in order to create more “natural” looking
flocking behavior.
Commented in deterring behavior that causes the fish to stay within the
confines of the fish tank.
Made initial fish movement random range greater (can travel in any
direction now).
Created constants for fish height and width for the deter function.