Added wall avoidance steering and corner escape behavior
Added behavior that steers a fish away from a wall if it is close
enough to the wall. Added since bouncing off of walls gets old.
The force that repels fish from walls is linearly related to the fish’s
distance from said wall.
There are rare scenarios where sharks can trap fish into corners, and
the fish are stuck in the corners because of their escape from shark
method forcing them into the corners. The fish thus spin in circles
while bouncing off the walls in the corners. This has been taken care
of by checking if a fish is caught in this exact situation (shark
nearby and forced in a corner), and telling the fish to travel in one
of two directions, the two directions being on each side of the shark.
This is calculated by a coin toss sort of method, and random floats
that determine the direction.
The solution to the corner problem is further taken care of in the
update method, where a counter is counted down for how long the fish is
to travel in the direction previously determined. The fish then resumes
its usual behaviors once the counter reaches 0.
A few functions in the fish class were reorganized to be a bit more
efficient. Specifically, the update method has been reorganized to
avoid calls where unnecessary.

Within the shark class, a separation method identical to the one for
fish was added.