Added Shark's eat function and Shark Wandering
Sharks will open their mouths if close enough to a fish. This function
simply uses distances between centers, and does not check if the fish
is in front of the shark or not (though the hunting algorithm makes it
such that this is most often the case).  Once the shark is close
enough, it bites down, and the fish is eaten, leaving behind a brief
particle simulation that is supposed to be a proxy for blood.

Once the shark has eaten, it wanders around the aquarium, switching
directions every second or so. A random direction is chosen and the
shark steers towards that direction. The shark stops wandering after 5
seconds or so and resumes feeding.