Added Fish add and Shark add buttons to UI, along with counters
Implemented via AquariumManager.cs

Also added failed implementations of binned structures for finding
neighbors. Idea was to use 2d grid and keep track of fish in each cell,
and pass the neighboring cells when finding neighbors. Could not find a
quick implementation of this. Used lists in each cell, and that took
even longer than the n^2 brute force implementation. Then used
dictionaries in each cell, and that was about equivalent to the n^2
implementation.