Figured out what was wrong!
Turns out that my "brute force" approach was checking for objects with
the tag "fish". Turns out that instantiating new gameobjects from a
prefab does not save the tag that the prefab had. Thus, when I was
adding new fish, the new fish weren't tagged as "Fish", and so they were
not being searched through within the brute force method. Thus the brute
force method was strangely close to linear time.

With ~300-400 fish, the brute force method achieves an FPS of around 20.

I have implemented 3 other methods (since back then I was still
wondering why the brute force was more efficient), and the 3 other
methods all grow at a much slower rate, albeit, some constant factors
are larger than others. The first method is through the use of a 2D List
with lists of fish in each cell.

The cells correspond to locations in the aquarium, and the list of fish
contained are the fish current in that boxed location in the aquarium.
In orer to find nearby fish, a fish simply queries it's own cell in the
grid, and neighboring cells, and then adds them to its neighbor list.
Then the calculations proceed as normal.
Note that this grid needs to be updated every frame based on a fish's
position. Additionally, deletions from a cell in the 2d array take
linear time, which isn't the biggest deal, but is still a good amount of
overhead.

The next method utilizes a 2D array but with a node to a doubly linked
list contained in each cell. The data of each node is a fish. Note that
since a node has fixed size, I can use an array here instead of a list.
By having pointers to their own nodes, each fish can delete itself from
the 2D array in constant time. They of course can also be added in
constant time to the array. However, doubly linked lists pose an issue
in the sense that they take a considerable amount of time to copy into a
list to be used by a fish when a fish searches for its neighbors. Thus,
this method had the largest constant factor, and was the slowest of the
three methods.

The quickest method involved the use of a 2d list of lists, of ints
(instead of fish in the first method). The ints contained in the list
were unique identifiers of the fish, and an array was used to store
pointers to the fish. Note that the unique identifiers of the fish could
be used as unique indexes into the array to access the fish they
correspond to. This method allowed for quicker updates to the 2D list
likely since the first method had lots of overhang due to managing
references to the fish in the 2D list rather than easy to work with ints
(which are stored contiguously in memory), although this is speculation.
This is the third and quickest method, and achieves an FPS of around
24-27 with 1,000 fish. Much faster than the brute force method, although
I am unclear as to the exact mathematical asymptotic runtime.