Cleaned up code
Added grid implementation for sharks, and generally cleaned up the
unnecessary code from prior experiments.

Additionally, I was not running the Boids algorithm before in order to
test out the speeds of fetching neighbors alone, and I commented Boids
in again. The FPS took a hit, around 400 fish yields around 20~30 fps,
depending on the density of the fish. I plan on implementing a k
neighbors algorithm in the next iteration in order to see how much of an
improvement it yields, and whether it maintains "good" flocking
behavior.