Added K-Nearest Neighbors Approach
Unfortunately, this method proved to be less efficient, and produce
worse flocking behavior compared to the simple grid access approach done
earlier.

The K-Nearest neighbors approach I created utilized the grid in order to
find the K-nearest neighbors to a fish within the neighboring grid
cells. I was hoping that by using K nearest neighbors, the reduction in
time for calculating the flock function (limiting the number of fish to
go through to K) would more than pay off the cost of finding the K
nearest neighbors. Unfortunately, this was not the case.

The above is due to the fact that in order to find the K-nearest
neighbors, I conducted a naive linear scan of all the fish in the local
area. I would find the minimum distance of all the fish, K times, to
find the K nearest neighbors. Turns out, what I was doing before
(calculating Boids on all the fish in the local area), was cheaper than
doing this when K=5. Since the old method could find greater than K=5
neighbors, I decided to use the old method instead.

I am considering using a KD tree to access the K nearest neighbors in K
* log(N) time where N is the number of fish. It is notable that
insertion into a KD tree takes roughly log(n) time, and since our data
(fish) change from frame to frame, each frame would have to update the
KD tree and thus this would result in N * log(N) runtime. Considering
that we must also run operations on k elements for each fish, the bound
would thus be O(N*log(N) + k * N), if we ignore dimensionality in
finding KNN.