Added KD Tree KNNs search!
The Distance class, ID class, and Node class are all helper classes for
the implementation of the KD tree. Distance and ID classes were created
simply to allow for the use of references to primitive data types in the
Nearest Neighbors function.

As stated in the KD Tree class, most of the insertion, deletion methods
of the class were created with the help of online sources. The Nearest
Neighbors function however was created through utilizing my recent
knowledge on how to search within KD-trees (though not extremely
complicated), and has been tailored for the use of KNNs search with
boids. The KD-tree functions like a binary search tree, and so the
insertion and deletion methods are similar to those found for a one
dimensional binary search tree, except tailored for the use of 2
dimensions (in my case).

Before doing this, I also tested a method to find the K nearest
neighbors by iterating through all the fish found by accessing the
neighboring cells in the 2D grid implementation. This proved to be
inefficient, even compared to the original 2D grid implementation. This
was because finding a neighbor proved to be an expensive search compared
to simply iterating through all the fish found in the neighboring cells
for the Boids algorithm.

The KD tree method is the fastest and most consistent of all the
algorithms I have tried so far,. Unlike the previous Grid
implementations, it does not fluctuate with the density of the fish.
With around 700 fish active, the KD tree implementation maintains 18-22
fps, while the Grid implementation (the second fastest) fluctuates
between 6-15 fps, and averages around 8 fps.

In terms of high level implementation of the KD tree, a new KD tree is
built once every update by the Aquarium Manager (and the old one
discarded). If the aquarium manager is updated after all the fish have
been updated, then for each frame, building a new KD tree takes roughly
n * log(n) time. The KD tree is built once per frame, inserts n fish
into a KD tree, where each insertion takes roughly log(n) time on
average. Then, for the updates to each fish, we must find K neighbors,
and search the KD tree, which takes roughly log(n) time. Thus, if we add
up the search time for all the fish, we get a search time proportional
to K * n * log(n). This is more efficient than the n * Q from the grid
implementation, where Q is the average amount of fish within the local
area, which grows proportionally to N, so once we hit large amounts of
fish, the grid implementation is roughly O(n^2). Since the update for
each fish is dominated by the search time, the run time for updated each
fish is asymptotic k*n*log(n) time (if each search takes roughly
log(n)).

I have also noted that the fps seems to slow down nearly linearly in the
number of fish, which is a good sign that the KD algorithm is working as
intended.

I have also implemented the KD algorithm for the sharks as well.