Added on Click dispersing behavior
Using 2D grid implementation to get fast access to the agents in a local
area. Thus, I am keeping track of a KD Tree and a 2D grid at the same
time. The 2D grid allows me to quickly access agents in a local area,
allowing me to get access to the agents in constant time, and then find
the agents within a certain distance in time proportional to the nunber
of agents. This allows me to search only a fraction of the total number
of fish for the fish that are close to the mouseclick.

The fish disperse in a fashion similar to the seperate component of the
boids algorithm. Their distance to the point of mouseclick corresponds
to how fast they will move away (or at least, the desired vector of
direction magnitude scales proportionally).

Fishes and sharks still take on their normal behavior when dispersing,
it is just that the disperse weight is large enough such that the other
effects are minimal in comparison. They disperse for 80 frames each.